netbox-ais-intergration/requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyYAML==5.4.1
requests==2.24.0
schedule==1.0.0
urllib3==1.25.11
Werkzeug==1.0.1